chromium: Force browser window size
Applications have no knowledge of expected surface size and the GDP HMI
stretches them to fill all the available space. According to
LayerManagerControl, the destination region for applications in the HMI
has a 1728x1080 size, so forcing the browser window to have that size
is a workaround to ensure the browser has the proper aspect ratio.

[GDP-477] Integrate Chromium on a Wayland/Ozone stack
[LM-2] Applications under ILM need to know expected surface size

Signed-off-by: Jacobo Aragunde Pérez <jaragunde@igalia.com>